length to unsigned int

+++ string_length.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int strlen(char *s){
+size_t strlen(char *s){
 	char *p = s;
 	int len = 0;
 	while(*p != '\0'){
@@ -14,7 +14,7 @@ int main(void){
 	int i;
 	char *s[] = {"Git tutorials", "Tutorials Point"};
 	for(i = 0; i < 2; i++ ){
-		printf("Length of %s is %d\n",s[i],strlen(s[i]));
+		printf("Length of %s is %lu\n",s[i],strlen(s[i]));
 	}
 
 }
